requirements.txt
django
djangorestframework
curveengine
flask
django-rest-swagger
drf-yasg